combine PathHitTest and PathHitTestManual into one test.
