--- test/jdk/java/awt/geom/HitTest/PathHitTestManual.java
@@ -1,257 +1,0 @@
-/*
- * Copyright (c) 1999, 2023, Oracle and/or its affiliates. All rights reserved.
- * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
- *
- * This code is free software; you can redistribute it and/or modify it
- * under the terms of the GNU General Public License version 2 only, as
- * published by the Free Software Foundation.
- *
- * This code is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
- * version 2 for more details (a copy is included in the LICENSE file that
- * accompanied this code).
- *
- * You should have received a copy of the GNU General Public License version
- * 2 along with this work; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
- *
- * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
- * or visit www.oracle.com if you need additional information or have any
- * questions.
- */
-
-import java.awt.BorderLayout;
-import java.awt.Canvas;
-import java.awt.Choice;
-import java.awt.Color;
-import java.awt.Dimension;
-import java.awt.EventQueue;
-import java.awt.Frame;
-import java.awt.Graphics;
-import java.awt.Graphics2D;
-import java.awt.Panel;
-import java.awt.Shape;
-
-/*
- *    @test
- *    @bug 4210936
- *    @summary Tests the results of the hit test methods on 3 different
- *           Shape objects - Polygon, Area, and GeneralPath.  Both an
- *           automatic test for constraint compliance and a manual
- *           test for correctness are included in this one class.
- *    @key headful
- *    @library /java/awt/regtesthelpers
- *    @build PassFailJFrame
- *    @run main/manual PathHitTestManual
- */
-
-public class PathHitTestManual extends Panel {
-    private static final String INSTRUCTIONS = """
-            This test displays the results of hit testing 5 different Shape
-            objects one at a time.
-
-            You can switch between shapes using the Choice component located
-            at the bottom of the window.
-
-            Each square in the test represents the
-            return values of the hit testing operators for that square region:
-
-                yellow - not yet tested
-                translucent blue overlay - the shape being tested
-
-                black - all outside
-                dark gray - rectangle intersects shape
-                light gray - rectangle intersects and center point is inside shape
-                white - rectangle is entirely contained in shape
-                red - some constraint was violated, including:
-                    rectangle is contained, but center point is not
-                    rectangle is contained, but rectangle.intersects is false
-                    centerpoint is contained, but rectangle.intersects is false
-
-            Visually inspect the results to see if they match the above table.
-            Note that it is not a violation for rectangles that are entirely
-            inside the path to be light gray instead of white since sometimes
-            the path is complex enough to make an exact determination expensive.
-            You might see this on the GeneralPath NonZero example where the
-            circles that make up the path cross over the interior of the shape
-            and cause the hit testing methods to guess that the rectangle is
-            not guaranteed to be contained within the shape.
-            """;
-
-    public static final int BOXSIZE = PathHitTest.BOXSIZE;
-    public static final int BOXCENTER = PathHitTest.BOXCENTER;
-    public static final int TESTSIZE = PathHitTest.TESTSIZE;
-
-    public static final int NUMTESTS = (TESTSIZE + BOXSIZE - 1) / BOXSIZE;
-
-    PathHitTestCanvas phtc;
-
-    public void init() {
-        setLayout(new BorderLayout());
-        phtc = new PathHitTestCanvas();
-        add("Center", phtc);
-        final Choice ch = new Choice();
-        for (int i = 0; i < PathHitTest.testNames.length; i++) {
-            ch.add(PathHitTest.testNames[i]);
-        }
-        ch.addItemListener(e -> phtc.setShape(ch.getSelectedIndex()));
-        ch.select(0);
-        phtc.setShape(0);
-        add("South", ch);
-    }
-
-    public void start() {
-        phtc.start();
-    }
-
-    public void stop() {
-        phtc.stop();
-    }
-
-    public static class PathHitTestCanvas extends Canvas implements Runnable {
-        public static final Color[] colors = {
-                /* contains?  point in?  intersects? */
-                Color.black,        /*    NO         NO          NO      */
-                Color.darkGray,     /*    NO         NO          YES     */
-                Color.red,          /*    NO         YES         NO      */
-                Color.lightGray,    /*    NO         YES         YES     */
-                Color.red,          /*    YES        NO          NO      */
-                Color.red,          /*    YES        NO          YES     */
-                Color.red,          /*    YES        YES         NO      */
-                Color.white,        /*    YES        YES         YES     */
-                Color.yellow,       /*     used for untested points      */
-        };
-
-        public Dimension getPreferredSize() {
-            return new Dimension(TESTSIZE, TESTSIZE);
-        }
-
-        public synchronized void start() {
-            if (!testdone) {
-                renderer = new Thread(this);
-                renderer.setPriority(Thread.MIN_PRIORITY);
-                renderer.start();
-            }
-        }
-
-        public synchronized void stop() {
-            renderer = null;
-        }
-
-        private Thread renderer;
-        private int shapeIndex = 0;
-        private byte[] indices = new byte[NUMTESTS * NUMTESTS];
-        boolean testdone = false;
-
-        private synchronized void setShape(int index) {
-            shapeIndex = index;
-            testdone = false;
-            start();
-        }
-
-        public void run() {
-            Thread me = Thread.currentThread();
-            Graphics2D g2d = (Graphics2D) getGraphics();
-            byte[] indices;
-            Shape s = PathHitTest.testShapes[shapeIndex];
-            synchronized (this) {
-                if (renderer != me) {
-                    return;
-                }
-                this.indices = new byte[NUMTESTS * NUMTESTS];
-                java.util.Arrays.fill(this.indices, (byte) 8);
-                indices = this.indices;
-            }
-
-            System.err.printf("%s %s\n", g2d, Color.yellow);
-            g2d.setColor(Color.yellow);
-            g2d.fillRect(0, 0, TESTSIZE, TESTSIZE);
-            int numtests = 0;
-            long start = System.currentTimeMillis();
-            for (int y = 0; renderer == me && y < TESTSIZE; y += BOXSIZE) {
-                for (int x = 0; renderer == me && x < TESTSIZE; x += BOXSIZE) {
-                    byte index = 0;
-                    if (s.intersects(x, y, BOXSIZE, BOXSIZE)) {
-                        index += 1;
-                    }
-                    if (s.contains(x+BOXCENTER, y+BOXCENTER)) {
-                        index += 2;
-                    }
-                    if (s.contains(x, y, BOXSIZE, BOXSIZE)) {
-                        index += 4;
-                    }
-                    numtests++;
-                    int i = (y / BOXSIZE) * NUMTESTS + (x / BOXSIZE);
-                    indices[i] = index;
-                    g2d.setColor(colors[index]);
-                    g2d.fillRect(x, y, BOXSIZE, BOXSIZE);
-                }
-            }
-            synchronized (this) {
-                if (renderer != me) {
-                    return;
-                }
-                g2d.setColor(new Color(0, 0, 1, .2f));
-                g2d.fill(s);
-                testdone = true;
-                long end = System.currentTimeMillis();
-                System.out.println(numtests+" tests took "+(end-start)+"ms");
-            }
-        }
-
-        public void paint(Graphics g) {
-            g.setColor(Color.yellow);
-            g.fillRect(0, 0, TESTSIZE, TESTSIZE);
-            byte[] indices = this.indices;
-            if (indices != null) {
-                for (int y = 0; y < TESTSIZE; y += BOXSIZE) {
-                    for (int x = 0; x < TESTSIZE; x += BOXSIZE) {
-                        int i = (y / BOXSIZE) * NUMTESTS + (x / BOXSIZE);
-                        g.setColor(colors[indices[i]]);
-                        g.fillRect(x, y, BOXSIZE, BOXSIZE);
-                    }
-                }
-            }
-            Graphics2D g2d = (Graphics2D) g;
-            g2d.setColor(new Color(0, 0, 1, .2f));
-            g2d.fill(PathHitTest.testShapes[shapeIndex]);
-        }
-    }
-
-
-    static volatile PathHitTestManual pathHitTestManual = null;
-    private static void createAndShowGUI() {
-        pathHitTestManual = new PathHitTestManual();
-        Frame frame = new Frame("PathHitTestManual test window");
-
-        frame.add(pathHitTestManual);
-        frame.setSize(400, 450);
-
-        PassFailJFrame.addTestWindow(frame);
-        PassFailJFrame.positionTestWindow(frame, PassFailJFrame.Position.HORIZONTAL);
-
-        frame.setVisible(true);
-
-        pathHitTestManual.init();
-        pathHitTestManual.start();
-    }
-
-    public static void main(String[] args) throws Exception {
-        PassFailJFrame passFailJFrame = new PassFailJFrame.Builder()
-                .title("PathHitTestManual Instructions")
-                .instructions(INSTRUCTIONS)
-                .testTimeOut(5)
-                .rows(30)
-                .columns(70)
-                .screenCapture()
-                .build();
-
-        EventQueue.invokeAndWait(PathHitTestManual::createAndShowGUI);
-        try {
-            passFailJFrame.awaitAndCheck();
-        } finally {
-            pathHitTestManual.stop();
-        }
-    }
-}
